# Fixed a race condition causing some workers to perform multiple tasks serially while other workers sleep.