@ben-clayton 2023-11-28
Firstly, my apologies for the delay in responding to this - it's been a very busy few weeks.

The extremely thorougher write-up is very much appreciated - excellent work!
The code changes also look good. I have no request for changes.

I did try benchmarking before and after this change, and I see mixed results:

```
go run ./tools/cmd/benchdiff/main.go old.txt new.txt

Delta               | Test name                                       | (A) old.txt  | (B) new.txt
+1.22x +26.989027ms | Schedule/Empty/tasks:262144/threads:8           | 123.310038ms | 150.299065ms
+1.20x +23.860334ms | Schedule/Empty/tasks:262144/threads:16          | 118.323588ms | 142.183922ms
+1.14x +15.550917ms | Schedule/Empty/tasks:262144/threads:128         | 107.884675ms | 123.435592ms
+1.14x +26.547827ms | Schedule/WaitGroup/tasks:262144/threads:8       | 189.861532ms | 216.409359ms
+1.14x +19.110006ms | Schedule/Empty/tasks:262144/threads:4           | 137.213622ms | 156.323628ms
+1.14x +23.777794ms | Schedule/WaitGroup/tasks:262144/threads:32      | 172.705833ms | 196.483627ms
+1.11x +12.103355ms | Schedule/Empty/tasks:262144/threads:32          | 114.35609ms  | 126.459445ms
+1.11x +72.549µs    | Schedule/Ticket/tasks:512/threads:1             | 687.316µs    | 759.865µs
+1.08x +8.820488ms  | Schedule/Empty/tasks:262144/threads:64          | 106.878291ms | 115.698779ms
+1.07x +14.902732ms | Schedule/WaitGroup/tasks:262144/threads:2       | 202.33975ms  | 217.242482ms
+1.07x +94.295µs    | Schedule/Event/tasks:512/threads:4              | 1.443245ms   | 1.53754ms
+1.06x +78.095µs    | Schedule/Event/tasks:512/threads:2              | 1.263009ms   | 1.341104ms
+1.05x +10.325158ms | Schedule/WaitGroup/tasks:262144/threads:4       | 205.738469ms | 216.063627ms
+1.04x +7.813932ms  | Schedule/WaitGroup/tasks:262144/threads:16      | 189.040758ms | 196.85469ms
+1.02x +2.051931ms  | Schedule/WaitGroup/tasks:262144/threads:1       | 133.807981ms | 135.859912ms
+1.02x +1.479942ms  | Schedule/EventBaton/tasks:262144/threads:32     | 96.728094ms  | 98.208036ms
+1.01x +15.027µs    | Schedule/Ticket/tasks:512/threads:128           | 1.06263ms    | 1.077657ms
+1.01x +2.318292ms  | Schedule/WaitGroup/tasks:262144/threads:128     | 171.273585ms | 173.591877ms
+1.01x +18.574µs    | Schedule/Event/tasks:512/threads:64             | 1.513371ms   | 1.531945ms
-1.01x -324ms       | Schedule/SomeWork/tasks:262144/threads:1        | 30.972s      | 30.648s
-1.01x -1.074193ms  | Schedule/EventBaton/tasks:262144/threads:64     | 96.596443ms  | 95.52225ms
-1.01x -4.38938ms   | MultiQueueTaskExecutor/tasks:262144/threads:128 | 387.045115ms | 382.655735ms
-1.01x -199ms       | Schedule/SomeWork/tasks:262144/threads:2        | 15.571s      | 15.372s
-1.01x -15.616µs    | Schedule/Ticket/tasks:512/threads:8             | 1.146066ms   | 1.13045ms
-1.02x -1.90551ms   | Schedule/EventBaton/tasks:262144/threads:4      | 97.675636ms  | 95.770126ms
-1.03x -47.992µs    | Schedule/Event/tasks:512/threads:32             | 1.568828ms   | 1.520836ms
-1.03x -406.474µs   | Schedule/Empty/tasks:262144/threads:0           | 12.695743ms  | 12.289269ms
-1.04x -53.122µs    | Schedule/Event/tasks:512/threads:128            | 1.531908ms   | 1.478786ms
-1.04x -3.788623ms  | Schedule/EventBaton/tasks:262144/threads:128    | 99.03754ms   | 95.248917ms
-1.05x -1.94971ms   | Schedule/Empty/tasks:262144/threads:1           | 39.77893ms   | 37.82922ms
-1.10x -93.617µs    | Schedule/Ticket/tasks:512/threads:64            | 1.05547ms    | 961.853µs
-1.11x -73.189µs    | Schedule/Event/tasks:512/threads:1              | 710.41µs     | 637.221µs
-1.12x -15.692897ms | Schedule/Empty/tasks:262144/threads:2           | 151.811635ms | 136.118738ms
-1.17x -205.383µs   | Schedule/Ticket/tasks:512/threads:2             | 1.43338ms    | 1.227997ms
-1.18x -25.829246ms | Schedule/WaitGroup/tasks:262144/threads:64      | 169.490816ms | 143.66157ms
```

This was run on Ubuntu 22.04.2, AMD Ryzen Threadripper 3990X.

There's a bit of a concerning performance loss with the `Schedule/Empty/tasks:*` benchmarks.

Do you see any delta with these benchmarks?